third_party/blink/web_tests/fast/events/tabindex-removal-from-focused-element-expected.txt

PASS document.activeElement is target
Remove tabindex.
PASS blur event was disaptched.
PASS successfullyParsed is true

TEST COMPLETE
